ENH: added background process support for posix
You can now run:

  chalmers start -w PROCESS

Then from any other window run:

  chalmers bg PROCESS

This will stop 'chalmers start' from printing to stdout/err and also ignore SIGHUB so you can close the window any time.

+++ chalmers/commands/log.py
@@ -33,8 +33,11 @@ def main(args):
         print(logfile)
         return
     if args.rm:
-        log.info("Removing log file '%s'" % logfile)
-        os.unlink(logfile)
+        if os.path.isfile(logfile):
+            log.info("Removing log file '%s'" % logfile)
+            os.unlink(logfile)
+        else:
+            log.warn("File '%s' does not exist" % logfile)
         return
 
     with open(logfile) as fd:
